[rpc] Fix electrum monitor parsing of long ints

+++ src/electrum/electrs.h
@@ -16,7 +16,7 @@ namespace electrum {
 std::string electrs_path();
 std::vector<std::string> electrs_args(int rpcport, const std::string& network);
 
-std::map<std::string, int> fetch_electrs_info();
+std::map<std::string, int64_t> fetch_electrs_info();
 
 } // ns electrum
